- (cleanup) moved routes to separate folder - (db) create image link field in User
